back_end/requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
attrs==21.2.0
black==21.6b0
click==8.0.1
coverage==5.5
ecdsa==0.17.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
pluggy==1.0.0
py==1.10.0
pyasn1==0.4.8
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-jose==3.3.0
pytz==2021.1
regex==2021.4.4
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Werkzeug==2.0.1
